requirements-test.txt
openwisp-utils[qa,selenium] @ https://github.com/openwisp/openwisp-utils/archive/refs/heads/1.3.tar.gz
django-redis~=6.0.0
mock-ssh-server~=0.9.1
responses~=0.25.8
psycopg2-binary~=2.9.10
pytest~=8.4.1
